backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
psycopg2-binary==2.9.10
pydantic-settings==2.9.1
PyJWT==2.10.1
pwdlib[argon2]==0.2.1
email-validator==2.2.0
pytest==8.3.5
httpx==0.28.1
